Revert "Move ui/pb package to ui/console/pb"
This reverts commit 8f8185832cdb35eb083546fa7a3b83ff47275766.
