+++ core/local/local.go
@@ -13,7 +13,7 @@ import (
 	"go.k6.io/k6/lib"
 	"go.k6.io/k6/lib/executor"
 	"go.k6.io/k6/metrics"
-	"go.k6.io/k6/ui/console/pb"
+	"go.k6.io/k6/ui/pb"
 )
 
 // ExecutionScheduler is the local implementation of lib.ExecutionScheduler
