+++ lib/executor/base_executor.go
@@ -9,7 +9,7 @@ import (
 
 	"go.k6.io/k6/lib"
 	"go.k6.io/k6/metrics"
-	"go.k6.io/k6/ui/console/pb"
+	"go.k6.io/k6/ui/pb"
 )
 
 // BaseExecutor is a helper struct that contains common properties and methods
