+++ lib/executor/constant_vus.go
@@ -12,7 +12,7 @@ import (
 	"go.k6.io/k6/lib"
 	"go.k6.io/k6/lib/types"
 	"go.k6.io/k6/metrics"
-	"go.k6.io/k6/ui/console/pb"
+	"go.k6.io/k6/ui/pb"
 )
 
 const constantVUsType = "constant-vus"
