--- ui/console/pb/doc.go
@@ -1,2 +1,0 @@
-// Package pb implements a progress bar component for CLI user interfaces.
-package pb
