+++ ui/pb/doc.go
@@ -1,0 +1,2 @@
+// Package pb implements a progressbar component for CLI user interfaces
+package pb
